inmatelocator/requirements.txt
django<1.9
django-localflavor
requests
requests-cache
lxml
jsonfield
python-levenshtein
fuzzywuzzy
usaddress
probablepeople
pytz
